Classwork Monday
Researched ways to import a csv document into a program. Played around with the Pandas library and havve some idea of how to import data from a csv file, but is still working my way to actually apply it to the project. Once I have this part done, the rest should be fairly easy.